Abstract coordinator server
