+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/operation/CancelJobOperation.java
@@ -33,7 +33,7 @@ public class CancelJobOperation extends AbstractJobAsyncOperation {
     @Override
     protected PassiveCompletableFuture<?> doRun() throws Exception {
         SeaTunnelServer service = getService();
-        return service.cancelJob(jobId);
+        return service.getCoordinatorServer().cancelJob(jobId);
     }
 
     @Override
