+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/operation/SubmitJobOperation.java
@@ -60,6 +60,6 @@ public class SubmitJobOperation extends AbstractJobAsyncOperation {
     @Override
     protected PassiveCompletableFuture<?> doRun() throws Exception {
         SeaTunnelServer seaTunnelServer = getService();
-        return seaTunnelServer.submitJob(jobId, jobImmutableInformation);
+        return seaTunnelServer.getCoordinatorServer().submitJob(jobId, jobImmutableInformation);
     }
 }
