+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/operation/TaskCompletedOperation.java
@@ -56,6 +56,7 @@ public class TaskCompletedOperation extends AsyncOperation {
     @Override
     protected PassiveCompletableFuture<?> doRun() throws Exception {
         CheckpointCoordinator checkpointCoordinator = ((SeaTunnelServer) getService())
+            .getCoordinatorServer()
             .getJobMaster(taskInfo.getJobId())
             .getCheckpointManager()
             .getCheckpointCoordinator(taskInfo.getPipelineId());
